pyproject.toml
[project]
name = "progetto-sde2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "altair>=5.5.0",
    "geopandas>=1.0.1",
    "polars-lts-cpu>=1.21.0",
    "streamlit>=1.41.1",
]
